+++ src/main/java/artifacts/common/init/ModComponents.java
@@ -1,7 +1,6 @@
 package artifacts.common.init;
 
 import artifacts.Artifacts;
-import artifacts.common.components.EntityKillTrackerComponent;
 import artifacts.common.components.SwimAbilityComponent;
 import artifacts.common.components.SyncedBooleanComponent;
 import dev.onyxstudios.cca.api.v3.component.ComponentKey;
@@ -17,13 +16,10 @@ public class ModComponents implements EntityComponentInitializer {
 			ComponentRegistryV3.INSTANCE.getOrCreate(Artifacts.id("dropped_item_entity"), SyncedBooleanComponent.class);
 	public static final ComponentKey<SwimAbilityComponent> SWIM_ABILITIES =
 			ComponentRegistryV3.INSTANCE.getOrCreate(Artifacts.id("swim_abilities"), SwimAbilityComponent.class);
-	public static final ComponentKey<EntityKillTrackerComponent> ENTITY_KILL_TRACKER =
-			ComponentRegistryV3.INSTANCE.getOrCreate(Artifacts.id("entity_kill_tracker"), EntityKillTrackerComponent.class);
 
 	@Override
 	public void registerEntityComponentFactories(EntityComponentFactoryRegistry registry) {
 		registry.registerFor(ItemEntity.class, DROPPED_ITEM_ENTITY, entity -> new SyncedBooleanComponent("wasDropped"));
 		registry.registerForPlayers(SWIM_ABILITIES, SwimAbilityComponent::new, RespawnCopyStrategy.LOSSLESS_ONLY);
-		registry.registerForPlayers(ENTITY_KILL_TRACKER, entity -> new EntityKillTrackerComponent(), RespawnCopyStrategy.LOSSLESS_ONLY);
 	}
 }